requirements/dev.txt

-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.2.9

# Testing
mypy==0.650
pytest==4.0.2
pytest-sugar==0.9.2
pytest-django==3.4.4
factory-boy==2.11.1

# Code quality
flake8==3.6.0